% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.basin.r
\name{basin.createAverage}
\alias{basin.createAverage}
\title{create Average from vector or matrix of basin (vector needs ng entries)}
\usage{
basin.createAverage(input, basinObject, Type = "land")
}
\arguments{
\item{input}{input vector of basin e.g. SoilwaterContent from basinObject}

\item{basinObject}{basinObject itself that belongs to input}

\item{Type}{as string (land or WB or all) to define which reference are should be used to calculated weighted mean as basin wide average (landfrac or 1-landfrac or 1) * GAREA}
}
\value{
float or vector
}
\description{
#Functions to get average values of basin
}
