% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.basin.r
\name{basin.createRaster}
\alias{basin.createRaster}
\title{create Raster from vector of basin}
\usage{
basin.createRaster(input, basinObject, spatialReference = T)
}
\arguments{
\item{input}{input vector of basin e.g. G_ALTITUDE from basinObject}

\item{basinObject}{basinObject itself that belongs to input (needed to create raster)}

\item{spatialReference}{if set to TRUE raster is returned, if set to FALSE matrix is returned}
}
\value{
raster or matrix with specified information (2D information of Basin)
}
\description{
Functions to deal with vectoried basin values in a nice and easy way
}
