% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.defBasin.r
\name{basin.defBasin}
\alias{basin.defBasin}
\title{Definition of Basin}
\usage{
basin.defBasin(basinObject)
}
\arguments{
\item{basinObject}{basinObject that needs to be updated with information}
}
\value{
changed basinObject
}
\description{
Function initiliazes basin (basin shape is defined and global information to build basin raster is saved)
}
