% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.getBasinInfo.r
\name{basin.getBasinInfo}
\alias{basin.getBasinInfo}
\title{Loading basin information from UNF-Files}
\usage{
basin.getBasinInfo(basinObject)
}
\arguments{
\item{basinObject}{for basin that needs to be updated}
}
\value{
updated basinObject
}
\description{
Function that loads info from files into basinObject
}
