% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.getGWfactor.r
\name{basin.getGWfactor}
\alias{basin.getGWfactor}
\title{Defining Groundwater Recharge Factor}
\usage{
basin.getGWfactor(textureInfo, slopeInfo, permaInfo, aqiferInfo)
}
\arguments{
\item{textureInfo}{- information from G_TEXTURE.UNF1}

\item{slopeInfo}{- information from G_SLOPE_CLASS.UNF1}

\item{permaInfo}{- information from G_PERMAGLAC.UNF1}

\item{aqiferInfo}{- information from G_AQ_FACTOR.UNF1}
}
\value{
groundwater factor to determine groundwater recharge
}
\description{
Function to prepare and get groundwater related parameter (gwFactor = aquiferFactor x permaFactor x slopeFactor x textureFactor)
}
