% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.getNeighbourcells.r
\name{basin.getNeighbourcells}
\alias{basin.getNeighbourcells}
\title{Definition of Neighbouring Cells}
\usage{
basin.getNeighbourcells(basinObject)
}
\arguments{
\item{basinObject}{basinObject that needs to be updated by informaiton of neighbouring cells}
}
\value{
updated basinObject
}
\description{
Function to determine neighbouringh cells which is a pre-requesite for routeorder
}
