% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.getRGmax.r
\name{basin.getRGmax}
\alias{basin.getRGmax}
\title{Definition of Rg_max}
\usage{
basin.getRGmax(textureInfo)
}
\arguments{
\item{textureInfo}{information from G_TEXTURE.UNF1}
}
\value{
RG_max as maximal groundwater recharge rate in [mm/d]
}
\description{
Function to determin maximal recharge rate for groundwater
}
