% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.getRoutingInfo.r
\name{basin.getRoutingInfo}
\alias{basin.getRoutingInfo}
\title{Loading Routing Information for Basin}
\usage{
basin.getRoutingInfo(basinObject)
}
\arguments{
\item{basinObject}{basinObject that needs to be updated with information}
}
\value{
updated basinObject
}
\description{
Function to read routing information in routing folder and update basinObject
}
