% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.loadingClimate.r
\name{basin.loadingClimate}
\alias{basin.loadingClimate}
\title{Loading climate data for basin}
\usage{
basin.loadingClimate(
  cont,
  grdc_number,
  basinIndex,
  transMatrix,
  simPeriodDate,
  climateShortcut,
  DataDir,
  G_WG3_WG2WITH5MIN,
  G_WG3_WATCH,
  typeName = "prec",
  readNewly = F,
  ClimateFormat
)
}
\arguments{
\item{cont}{continent as string (au, af, as, eu, na, sa)}

\item{grdc_number}{grdc_number of basin (is used as ID in this case, however, later on it is also used to load Discharge data)}

\item{basinIndex}{basinIndex information to determine which cells belong to the basin}

\item{transMatrix}{transformation Matrix to transform global data to basinwide data (not sure if this is correct)}

\item{simPeriodDate}{DateVector defining the simulation period}

\item{climateShortcut}{filepath to climateshortcut-folder - usually ...base/data/ClimateShortcut}

\item{DataDir}{filepath to data-folder - usually ...base/data}

\item{G_WG3_WG2WITH5MIN}{transformation information to get basinwide information from global met. data}

\item{G_WG3_WATCH}{transformation information to get basinwide information from global met. data}

\item{typeName}{of variable to be read}

\item{readNewly}{information if climate data is forced to be read newly (T) or not (F, ClimateShortCut will be used)}

\item{ClimateFormat}{defining format of climate (global, continental or basin)}
}
\value{
Matrix with climate data (nrow=days, ncol=cells of basin)
}
\description{
Function to load climate when initiliazing climate object (calls other loadingClimate-functions, depending on climateFormat)
}
