% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.loadingClimate4Basin.R
\name{basin.loadingClimate4Basin}
\alias{basin.loadingClimate4Basin}
\title{Loading basinwide climate data}
\usage{
basin.loadingClimate4Basin(
  name = "prec",
  grdc_number,
  basIndex,
  DataDir,
  simPeriodDate
)
}
\arguments{
\item{name}{name of variable to be read}

\item{grdc_number}{grdc number of basin that is used to save climate data in ClimateShortcut-folder}

\item{basIndex}{basinIndex to define which cells of continental raster belong to basin}

\item{DataDir}{filepath to data-folder - usually ...base/data}

\item{simPeriodDate}{DateVector defining the simulation period}
}
\value{
Matrix with climate data (nrow=days, ncol=cells of basin)
}
\description{
Function to read basinwide climate data
}
