% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.loadingClimate4Cont.r
\name{basin.loadingClimate4Cont}
\alias{basin.loadingClimate4Cont}
\title{Loading climate from continental data}
\usage{
basin.loadingClimate4Cont(
  name = "prec",
  basIndex,
  transMatrix,
  DataDir,
  simPeriodDate,
  cont
)
}
\arguments{
\item{name}{name name of variable to be read}

\item{basIndex}{basinIndex to define which cells of continental raster belong to basin}

\item{transMatrix}{transformation Matrix to transform global data to basinwide data (not sure if this is correct)}

\item{DataDir}{filepath to data-folder - usually ...base/data}

\item{simPeriodDate}{DateVector defining the simulation period}

\item{cont}{continent as string (au, af, as, eu, na, sa)}
}
\value{
Matrix with climate data (nrow=days, ncol=cells of basin)
}
\description{
Function to read continental climate infromation
}
