% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.loadingClimateEWEMBI.r
\name{basin.loadingClimateEWEMBI}
\alias{basin.loadingClimateEWEMBI}
\title{Loading worldwide EWEMBI data}
\usage{
basin.loadingClimateEWEMBI(
  name = "prec",
  basIndex,
  transMatrix,
  DataDir,
  G_WG3_WG2WITH5MIN,
  G_WG3_WATCH,
  simPeriodDate
)
}
\arguments{
\item{name}{name of variable to be read}

\item{basIndex}{basinIndex to define which cells of continental raster belong to basin}

\item{transMatrix}{transformation Matrix to transform global data to basinwide data (not sure if this is correct)}

\item{DataDir}{filepath to data-folder - usually ...base/data}

\item{G_WG3_WG2WITH5MIN}{transformation information to get basinwide information from global met. data}

\item{G_WG3_WATCH}{transformation information to get basinwide information from global met. data}

\item{simPeriodDate}{DateVector defining the simulation period}
}
\value{
Matrix with climate data (nrow=days, ncol=cells of basin)
}
\description{
Function to read worldwide (EWEMBI) climate data for basin
}
