% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.prepareRun.r
\name{basin.prepareRun}
\alias{basin.prepareRun}
\title{Preparation of model run (in rcpp)}
\usage{
basin.prepareRun(basinObject, climateObject, waterUseObject)
}
\arguments{
\item{basinObject}{basinObject defined from init.model()}

\item{climateObject}{climateObject defined from init.climate()}

\item{waterUseObject}{waterUseObject defined from init.waterUse()}
}
\value{
List to pass to rcpp function to run model
}
\description{
Function to create List to pass to WGL
}
