% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibratio.calibrateModel.r
\name{calibration.calibrateModel}
\alias{calibration.calibrateModel}
\title{Function to calibrate basin}
\usage{
calibration.calibrateModel(
  BasinObject2Cal,
  List2use4Cal,
  Settings,
  nWarmUp = 5,
  startVal = c(2.5),
  upperVal = c(0.1),
  lowerVal = c(5)
)
}
\arguments{
\item{BasinObject2Cal}{BasinObject to be calibrated}

\item{List2use4Cal}{Model Input for Basin}

\item{Settings}{Vector for settings to define which model setting should be used to run the model}

\item{nWarmUp}{number of years that should be used as warm-up, i.e. that should not be considered in model run}

\item{startVal}{gamma value to start calibration, e.g. c(2.5)}

\item{upperVal}{upper value for gamma, e.g. c(5.0)}

\item{lowerVal}{lower value gamma, e.g. c(0.1)}
}
\value{
information about calibration result
}
\description{
gamma is changed and gamma value is searched for which mean deviation between simulated and observed discharge is minimal
}
