% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.changeVars.r
\name{calibration.changeVars}
\alias{calibration.changeVars}
\title{Function to change model parameters with certain values}
\usage{
calibration.changeVars(ListConst, X)
}
\arguments{
\item{ListConst}{input List for model run}

\item{X}{Vector that contains value for gamma to be changed, e.g. c(2.5)}
}
\value{
Changed List that can be used for new model run with different gamma value
}
\description{
At the moment only Gamma is changed
}
