% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createWaterBalance}
\alias{createWaterBalance}
\title{Calculating waterbalance of basin}
\usage{
createWaterBalance(timestring)
}
\arguments{
\item{timestring}{Datevector with dates of simulation period}
}
\value{
List Vwith daily water balance for whole simulaiton period as output
}
\description{
{
daily routine for each cell
this routine is already validated against WG3 simulation for a small basin in South America
especially the snow routine and immediate RunOff generation are NOT validated therefore 
have to check the simulation results also for other basins to assure a good match with WG3
e.g. for european basins with snow processes and sealed areas! (Bayern?) }
}
