% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dailyEstimateLongwave}
\alias{dailyEstimateLongwave}
\title{Calculate longwave radiation}
\usage{
dailyEstimateLongwave(n, DOY, dailyTempC, dailyShortWave)
}
\arguments{
\item{n}{index of specific cell}

\item{DOY}{day of the year (between 1 and 365)}

\item{dailyTempC}{Temperature of Day in Degree}

\item{dailyShortWave}{shortwave radiation as double in W/m²}
}
\value{
net_long_wave_rad net longwave radiation in W/m²
}
\description{
estimation of longwave radiation when it is not given as measured variable
}
