% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dailyEstimateShortwave}
\alias{dailyEstimateShortwave}
\title{Calculate shortwave radiation (after Kaspar 2004)}
\usage{
dailyEstimateShortwave(
  dates_of_simulation,
  temperature,
  sunshine_duration,
  row_information_GR,
  corrected_row_for_continent
)
}
\arguments{
\item{dates_of_simulation}{Datevector of Simulation period}

\item{temperature}{Temperatur as NumericMatrix in degree}

\item{sunshine_duration}{Sunshine duration duration as NumericMatrix  in hours}

\item{row_information_GR}{information of row for cells}

\item{corrected_row_for_continent}{information of correction of rows for continental grid}
}
\value{
shortwave_downward_radiation Matrix with estimated shortwave radiation in W/m²
}
\description{
rcpp function to estimate Shortwave radiation when not given as measured input
}
