% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dailySnow}
\alias{dailySnow}
\title{snow storage interpolation}
\usage{
dailySnow(
  day,
  daily_prec_to_soil,
  G_snow,
  G_snowWaterEquivalent,
  dailySnowMelt,
  dailySnowEvapo,
  thresh_elev,
  dailyEffPrec,
  dailySoilPET
)
}
\arguments{
\item{day}{day of simulation period as integer}

\item{daily_prec_to_soil}{throughfall from canopy}

\item{G_snow}{snow storage at 5min scale}

\item{G_snowWaterEquivalent}{amount of water in snow storage at sub-grid scale}

\item{dailySnowMelt}{snow melt on 5min cell}

\item{dailySnowEvapo}{sublimation of snow on 5min cell}

\item{thresh_elev}{helper - information of reference height,when there is unlimited snow accummulation (> 1000mm)}

\item{dailyEffPrec}{effective precipitation to soil (throughfall + snow melt - fallen snow)}

\item{dailySoilPET}{energy for PET which is left for soil}
}
\description{
snow storage is calculated in sub-grid scale (1min) and aggregated to 5min after each day/iteration
}
