% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findNumberInVector}
\alias{findNumberInVector}
\title{findNumberInVector}
\usage{
findNumberInVector(number, vec)
}
\arguments{
\item{number}{integer to search for}

\item{vec}{Integervector that is searched in}
}
\value{
indices of number in vector (0=1st entry of vector)
}
\description{
Function that finds number in vector and return indices
}
