% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findUniqueValues}
\alias{findUniqueValues}
\title{findUniqueValues}
\usage{
findUniqueValues(vec)
}
\arguments{
\item{vec}{Integervector that is examined}
}
\value{
Vector with values in vector without duplicates
}
\description{
Function that returns values in vector without duplicates
}
