% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.getFileList.r
\name{getFileList}
\alias{getFileList}
\title{Defining Files to be read into model}
\usage{
getFileList(Type = "basinInfo")
}
\arguments{
\item{Type}{that defines which information is required (basinInfo, changingFiles or routing)}
}
\value{
FileList with files that needs to be read
}
\description{
Function to get FileList to read
}
