% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.getInfoFile.r
\name{getInfoFile}
\alias{getInfoFile}
\title{Getting File Information of UNF-File}
\usage{
getInfoFile(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\value{
list(nlayers, Size, Type)
}
\description{
gets File info how to decode binaries
}
