% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRiverVelocity}
\alias{getRiverVelocity}
\title{getRiverVelocity}
\usage{
getRiverVelocity(Type, cell, inflow)
}
\arguments{
\item{Type}{0 (constant) or 1 (variable)}

\item{cell}{cell that is simulated}

\item{inflow}{inflow to river in mm*km²/day}
}
\value{
riverVelocity in km/day
}
\description{
function that defines river velocity for routing (variable or constant)
}
