% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.basin.r
\name{init.basin}
\alias{init.basin}
\title{Initializing Basin as empty basinObject}
\usage{
init.basin(continent_object, grdc_number, lat, long, cont)
}
\arguments{
\item{continent_object}{ContinentObject defined from init.initCont()}

\item{grdc_number}{grdc number of basin}

\item{lat}{latitude of basin outlet in degree}

\item{long}{longitud of basin outlet in degree}

\item{cont}{continent as string (au, as, af, eu, na, sa)}
}
\value{
created basinObject
}
\description{
function to initialize basin
}
