% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.gettingPaths.r
\name{init.gettingPaths}
\alias{init.gettingPaths}
\title{Path Definition}
\usage{
init.gettingPaths(cont, base = NULL)
}
\arguments{
\item{cont}{continent as string where basin is part of (as, af, au, na, eu, sa)}

\item{base}{filepath to folder where folder for data, source_code and output are located}
}
\value{
named list with pathes to use
}
\description{
Function to determine paths in r so package can find input and knows where to write the output
}
