% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.initCont.r
\name{init.initCont}
\alias{init.initCont}
\title{Initializing continentalObject}
\usage{
init.initCont(cont, path2base)
}
\arguments{
\item{cont}{continent as string where basin is part of (as, af, au, na, eu, sa)}

\item{path2base}{filepath to folder where folder for data, source_code and output are located}
}
\value{
new continent object for continent where basin is located
}
\description{
Function to initialize the continent where basin is located
}
