% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.setParsBasin.r
\name{init.SetParsBasin}
\alias{init.SetParsBasin}
\title{Definition of model parameter values for basinObject}
\usage{
init.SetParsBasin(basinObject, LAI_info, LCT_info)
}
\arguments{
\item{basinObject}{that is updated with parameters}

\item{LAI_info}{table with LAI_info (is then used to define parameters)}

\item{LCT_info}{table with LCT info (is then used to define parameters)}
}
\value{
basin object with defined parameters  for basin (continent object is part of basin object)
}
\description{
Function to set parameters of model to basin
}
