% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.waterUse.r
\name{init.waterUse}
\alias{init.waterUse}
\title{Initializing water use information for model run}
\usage{
init.waterUse(basinObject, simStart, simEnd, WaterUse_Setting = 0)
}
\arguments{
\item{basinObject}{that defines basin for which water use is loaded}

\item{simStart}{as string of Type \%dd.\%mm.\%YYYY, e.g. 01.01.1980}

\item{simEnd}{as string of Type \%dd.\%mm.\%YYYY, e.g. 01.01.1990}

\item{WaterUse_Setting}{0=noWaterUse is considered and dummy with 0 is loaded, != 1 water use is loaded}
}
\value{
new waterUse object for basin
}
\description{
function to initialize water uses
}
