% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numberOfDaysInMonth}
\alias{numberOfDaysInMonth}
\title{numberOfDaysInMonth}
\usage{
numberOfDaysInMonth(month, year)
}
\arguments{
\item{month}{as integer (1 = january)}

\item{year}{as integer}
}
\value{
number of days of specified month in specified year as integer
}
\description{
Function that gives the number of Days in month
}
