% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numberOfDaysInYear}
\alias{numberOfDaysInYear}
\title{numberOfDaysInYear}
\usage{
numberOfDaysInYear(year)
}
\arguments{
\item{year}{as integer}
}
\value{
number of days of specified year as integer
}
\description{
Function that gives the number of Days in year
}
