% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.Q.r
\name{Q.calcQuality}
\alias{Q.calcQuality}
\title{Calculating Benchmark}
\usage{
Q.calcQuality(df_obs, df_sim, Type = "NSE")
}
\arguments{
\item{df_obs}{data.frame of observed discharge with ("Date" and "Value" columns)}

\item{df_sim}{data.frame of simulated discharge with ("Date" and "Sim" Columns)}

\item{Type}{string (QmeanAbs, NSE, logNSE, KGE or SFDCE)}
}
\value{
named list with benchmarks
}
\description{
Function to calculate several benchmarks to compare simulated and observed discharge. NA or negative values in observed data are omitted.
}
