% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.Q.r
\name{Q.convert_m3s_mmday}
\alias{Q.convert_m3s_mmday}
\title{m³/s to mm/d}
\usage{
Q.convert_m3s_mmday(timeseries, area)
}
\arguments{
\item{timeseries}{vector/matrix of data that needs to be converted}

\item{area}{sum of GAREA information of basin, (area must be in km²)}
}
\value{
changed timeseries
}
\description{
Functions to convert m³/s to mm/d
}
