% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.Q.r
\name{Q.convert_mmday_m3s}
\alias{Q.convert_mmday_m3s}
\title{mm/d to m³/s}
\usage{
Q.convert_mmday_m3s(timeseries, area)
}
\arguments{
\item{timeseries}{vector/matrix of data that needs to be converted}

\item{area}{GAREA information of basin, can be obtained grom GAREA of basinObject}
}
\value{
changed timeseries
}
\description{
Functions to convertmm/d to m³/s
}
