% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.Q.r
\name{Q.createMonthlyPlot}
\alias{Q.createMonthlyPlot}
\title{Plotting monthly timeseries}
\usage{
Q.createMonthlyPlot(df_obs, df_sim)
}
\arguments{
\item{df_obs}{data.frame of observed daily discharge with ("Date" and "Value" columns)}

\item{df_sim}{data.frame of simulated daily discharge with ("Date" and "Sim" Columns)}
}
\value{
ggplot object
}
\description{
Function to plot monthly timeseries (aggregation is done function intern)
}
