% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.Q.r
\name{Q.plotTimeseries}
\alias{Q.plotTimeseries}
\title{Plotting daily timeseries}
\usage{
Q.plotTimeseries(df_obs, df_sim, df_prec, from = NULL, to = NULL, showPlot = T)
}
\arguments{
\item{df_obs}{data.frame of observed daily discharge with ("Date" and "Value" columns)}

\item{df_sim}{data.frame of simulated daily discharge with ("Date" and "Sim" Columns)}

\item{df_prec}{data.frame of daily precipitation as basinwide average ("Date" and "prec" Columns)}

\item{from}{optional information (Date) where to start plot}

\item{to}{optional information (Date) where to end plot}

\item{showPlot}{if TRUE than plot is directly shown in Viewer, inf FALSE than not}
}
\value{
ggplot object
}
\description{
Functions to plot nicely results
}
