% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.Q.r
\name{Q.readGRDC}
\alias{Q.readGRDC}
\title{Reading GRDC data}
\usage{
Q.readGRDC(
  basinObject_id,
  basinObject_DataDir,
  basinObject_cont,
  start,
  end,
  useFolder = NULL
)
}
\arguments{
\item{basinObject_id}{grdc_number of basin, can be obtained from basinObject}

\item{basinObject_DataDir}{dataDirection as string can be obtained from basinObject}

\item{basinObject_cont}{continent as string (au, af, as, na, eu, sa)}

\item{start}{start data as date to read GRDC data}

\item{end}{end date as date to read GRDC data}

\item{useFolder}{if defined another folder than data/calibration/cont can be set to look for discharge data (GRDC-like data)}
}
\value{
dataframe with Date and Q as given in GRDC-file with name "(grdc_number)_Q_Day.Cmd.txt"
}
\description{
Functions to read discharge data
}
