% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.readingUNF.r
\name{readingUNF}
\alias{readingUNF}
\title{Reading UNF-File}
\usage{
readingUNF(
  file2read,
  DataDir,
  transMatrix = NULL,
  basIndex,
  name = "basinInfo",
  ng_land = NULL,
  TypeClimate = 0,
  cont = NULL
)
}
\arguments{
\item{file2read}{filename to read}

\item{DataDir}{defined (global) direction for data - usually workindgirectory/data/}

\item{transMatrix}{usually gcrc information}

\item{basIndex}{vector to use to define basin}

\item{name}{name folder to use for reading}

\item{ng_land}{ng_land information, needed to read G_ELEV_RANGE.UNF}

\item{TypeClimate}{value to determine which type needs to be written: 0: normal layer, 1:global climate data 2:continental climate data}

\item{cont}{character to determine continent (af,au,as,eu,na,sa)}
}
\value{
vector (single layer) oder matrix (mutlple layer) with each column is information for one cell in defined basin
}
\description{
function to read data from UNF-File
}
