% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{routing}
\alias{routing}
\title{routing}
\usage{
routing(SimPeriod, surfaceRunoff, GroundwaterRunoff, PETw, Prec)
}
\arguments{
\item{SimPeriod}{Datevector of Simulationperiod}

\item{surfaceRunoff}{run-off from surface contributing to river network}

\item{GroundwaterRunoff}{run-off from groundwater contributing to river network}

\item{PETw}{Potential Evapotranspiration as NumericMatrix in mm/d (to calculate water balance of waterbodies)}

\item{Prec}{Precipitation as NumericMatrix in mm/d}
}
\description{
this function includes als routing modules
}
