% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{routingRiver}
\alias{routingRiver}
\title{routingRiver}
\usage{
routingRiver(cell, riverVelocity, RiverInflow, G_riverOutflow, S_river)
}
\arguments{
\item{cell}{cell that is simulated}

\item{riverVelocity}{river velocity in km/d}

\item{RiverInflow}{inflow to river network [mm*km²/d]}

\item{G_riverOutflow}{transportedVolume in [mm*km²/d]}

\item{S_river}{river storage [mm*km²]}
}
\value{
transportedVolume in [mm*km²/d]
}
\description{
function that defines routing through river - note: uses original model code with bug in ELS equation
}
