% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(SimPeriod, ListConst, Settings, nYears)
}
\arguments{
\item{SimPeriod}{Period to simulate (usually defines as model[["SimPeriod"]] where model is object returned from basin.prepareRun())}

\item{ListConst}{list with all required information regarding basin and input (usually, object returned by basin.prepareRun())}

\item{Settings}{vector that is used to define settings}

\item{nYears}{number of years defined as warm-up (first year is then simulated n times, before starting with the actual simulation period)}
}
\description{
run whole model including model initializing, warm-up period, water balance and routing and returns list with states and fluxes
}
