% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setLakeWetlandToMaximum}
\alias{setLakeWetlandToMaximum}
\title{setLakeWetlandToMaximum}
\usage{
setLakeWetlandToMaximum(
  S_locLakeStorage,
  S_locWetlandStorage,
  S_gloLakeStorage,
  S_ResStorage,
  S_gloWetlandStorage
)
}
\arguments{
\item{S_locLakeStorage}{local lake storage}

\item{S_locWetlandStorage}{local wetland storage}

\item{S_gloLakeStorage}{global lake storage}

\item{S_ResStorage}{reservoir storage}

\item{S_gloWetlandStorage}{global wetland storage}
}
\description{
function that sets storage of all surface water bodies to max; is implemented in original model version (actually because of this, warm-up-period should be quite long > 2a)
}
