% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sortIt}
\alias{sortIt}
\title{sortIt}
\usage{
sortIt(vec)
}
\arguments{
\item{vec}{Integervector that is examined}
}
\value{
sorted vector
}
\description{
Function that sort a vector in ascending order (or descending not sure about that)
}
