% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumVector}
\alias{sumVector}
\title{sumVector}
\usage{
sumVector(vec)
}
\arguments{
\item{vec}{Numericvector that is sumed up}
}
\value{
sum of vector as double
}
\description{
Function that sums up a vector
}
