% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.prepareFolderStructure.r
\name{tools.prepareFolderStructur}
\alias{tools.prepareFolderStructur}
\title{tools.prepareFolderStructur}
\usage{
tools.prepareFolderStructur(base)
}
\arguments{
\item{base}{filepath that is used as base to prepare folder structure}
}
\description{
prepare FolderStructure to prepare model run with package
}
