% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tools_DefDrainageCells}
\alias{tools_DefDrainageCells}
\title{tools_DefDrainageCells}
\usage{
tools_DefDrainageCells(Outlet, GCRC, OutflowMatrix)
}
\arguments{
\item{Outlet}{of basin as GCRC number in continental grid}

\item{GCRC}{as vector for continent}

\item{OutflowMatrix}{as vector for continent}
}
\value{
IntegerVector with GCRC-IDs that are part of basin
}
\description{
rcpp tool to define Drainage Cells
}
