% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tools_InterpolateValues}
\alias{tools_InterpolateValues}
\title{tools_InterpolateValues}
\usage{
tools_InterpolateValues(VectorIn, Table_x, Table_y)
}
\arguments{
\item{VectorIn}{vector as base for interpolation}

\item{Table_x}{Table_x of values}

\item{Table_y}{Table_y of values}
}
\value{
NumericVector with interpolated output
}
\description{
rcpp function for interpolation of a vector (alternative could be using r function and apply)
}
