% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterUseCalcDaily}
\alias{WaterUseCalcDaily}
\title{WaterUseCalcDaily}
\usage{
WaterUseCalcDaily(
  waterUseType,
  dailyUse,
  year,
  month,
  StartYear,
  Info_GW,
  Info_SW,
  Info_TF
)
}
\arguments{
\item{waterUseType}{0 (no water use), 1 (only water use without Transport to cities) or 2 (only water use with Transport to cities)}

\item{dailyUse}{Matrix with two rows that gives water use for actual day in mm*km²/day (first = GW, second=SW+TF), note that all days in one month in one year have same values}

\item{year}{year of simulation period as integer}

\item{month}{month of simulation period as integer}

\item{StartYear}{information to get the right column from read water use information}

\item{Info_GW}{read water use information from groundwater}

\item{Info_SW}{read water use information from surface water}

\item{Info_TF}{read water use information for transport to cities}
}
\description{
calculates daily water use for groundwater and surface water (note that GapYearType needs to be included in function input)
}
