% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterUseCalcMeanDemandDaily}
\alias{WaterUseCalcMeanDemandDaily}
\title{WaterUseCalcMeanDemandDaily}
\usage{
WaterUseCalcMeanDemandDaily(year, GapYearType)
}
\arguments{
\item{year}{year of simulation period as integer}

\item{GapYearType}{Info from Setting wheter 29.02 is simulated (0) or not (1)}
}
\value{
G_mean_demand as numericVector for the sepcified year in [mm*km²/day]
}
\description{
Prepare infromation for reservoir (calculate yearly mean demand of cell itself and next 20 downstream cells)
}
