% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WaterUseConsumGW}
\alias{WaterUseConsumGW}
\title{WaterUseConsumGW}
\usage{
WaterUseConsumGW(cell, GroundwaterStorage, dailyUse)
}
\arguments{
\item{cell}{cell in basin that is used for abstraction}

\item{GroundwaterStorage}{groundwater storage level (can be negative due to abstraction)}

\item{dailyUse}{information of water that needs to be abstracted from groundwater (first row of NumericMatrix)}
}
\value{
GWdailyuse abstracted groundwater (if there is no landfraction in cell, no water can be abstracted from groundwater)
}
\description{
function that abstracts water use from groundwater storage
}
