% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.loadingClimate.r
\name{basin.loadingClimate}
\alias{basin.loadingClimate}
\title{Loading climate data for basin}
\usage{
basin.loadingClimate(
  cont,
  grdc_number,
  basin_index,
  trans_matrix,
  sim_period_date,
  climate_shortcut,
  data_dir,
  wg2with5min_mask,
  watch_mask,
  type_name = "prec",
  read_newly = FALSE,
  climate_format
)
}
\arguments{
\item{cont}{continent as string (au, af, as, eu, na, sa)}

\item{grdc_number}{grdc_number of basin (is used as ID in this case, however, later on it is also used to load Discharge data)}

\item{basin_index}{basinIndex information to determine which cells belong to the basin}

\item{trans_matrix}{transformation Matrix to transform global data to basinwide data (not sure if this is correct)}

\item{sim_period_date}{DateVector defining the simulation period}

\item{climate_shortcut}{filepath to climateshortcut-folder - usually ...base/data/ClimateShortcut}

\item{data_dir}{filepath to data-folder - usually ...base/data}

\item{wg2with5min_mask}{transformation information to get basinwide information from global met. data}

\item{watch_mask}{transformation information to get basinwide information from global met. data}

\item{type_name}{of variable to be read}

\item{read_newly}{information if climate data is forced to be read newly (T) or not (F, ClimateShortCut will be used)}

\item{climate_format}{defining format of climate (global, continental or basin)}
}
\value{
Matrix with climate data (nrow=days, ncol=cells of basin)
}
\description{
Function to load climate when initiliazing climate object (calls other loadingClimate-functions, depending on climateFormat)
}
