% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.waterUse.r
\name{init.wateruse}
\alias{init.wateruse}
\title{Initializing water use information for model run}
\usage{
init.wateruse(basin_object, sim_start, sim_end, wateruse_setting = 0)
}
\arguments{
\item{basin_object}{that defines basin for which water use is loaded}

\item{sim_start}{as string of Type \%dd.\%mm.\%YYYY, e.g. 01.01.1980}

\item{sim_end}{as string of Type \%dd.\%mm.\%YYYY, e.g. 01.01.1990}

\item{wateruse_setting}{0=noWaterUse is considered
and dummy with 0 is loaded, != 1 water use is loaded}
}
\value{
new waterUse object for basin
}
\description{
function to initialize water uses
}
