% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(SimPeriod, ListConst, Settings, nYears)
}
\arguments{
\item{SimPeriod}{Period to simulate (usually defines as model[["SimPeriod"]] where model is object returned from basin.prepareRun())}

\item{ListConst}{list with all required information regarding basin and input (usually, object returned by basin.prepareRun())}

\item{Settings}{vector of length 8 that is used to define settings:
\itemize{
 \item 1st entry: water use               -> 0 (off), 1 (on), 2 (on, including water transport to cities)
 \item 2nd entry: water use allocation    -> 0 (temporal & spatial distr.), 1 (spatial distr.), 2 (temporal distr.)
 \item 3rd entry: flow velocity           -> 0 (constant), 1 (variable)
 \item 4th entry: gap year                -> 0 (including 29.02), 1 (without 29.02)
 \item 5th entry: reservoir algorithm     -> 0 (Hanasaki), 1 (global lake)
 \item 6th entry: splitting factor        -> 0 (original version), 1 (set as parameter) - only used for development purposes
 \item 7th entry: long wave radiation     -> 0 (reading), 1 (calculating)
 \item 8th entry: warm up period          -> 0 (no system values), 1 (system values are read), 2 (system values are written), 3 (system values are read and written) }}

\item{nYears}{number of years defined as warm-up (first year is then simulated n times, before starting with the actual simulation)}
}
\description{
run whole model including model initializing, warm-up period, water balance and routing and returns list with states and fluxes
}
